// src/services/openapi/models/DailyAnalyticOutput.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type DailyAnalyticOutput = {
    id: string;
    active_links: number;
    day_links: number;
    day_openning: number;
    nb_differents_email_send: number;
    reference_date: string;
    created_at: string;
};